Convert attributes to 64-bit long type

+++ SmartInsiderEvent.cs
@@ -64,7 +64,7 @@ namespace QuantConnect.DataSource
         /// <summary>
         /// Smart Insider proprietary identifier for the company
         /// </summary>
-        public int? CompanyID { get; set; }
+        public long? CompanyID { get; set; }
 
         /// <summary>
         /// FTSE Russell Sector Classification
@@ -89,7 +89,7 @@ namespace QuantConnect.DataSource
         /// <summary>
         /// Numeric code that is the most granular level in ICB classification
         /// </summary>
-        public int? ICBCode { get; set; }
+        public long? ICBCode { get; set; }
 
         /// <summary>
         /// Company name. PLC is always excluded
@@ -230,12 +230,12 @@ namespace QuantConnect.DataSource
             LastIDsUpdate = string.IsNullOrWhiteSpace(tsv[4]) ? (DateTime?)null : DateTime.ParseExact(tsv[4], "yyyyMMdd", CultureInfo.InvariantCulture);
             ISIN = string.IsNullOrWhiteSpace(tsv[5]) ? null : tsv[5];
             USDMarketCap = string.IsNullOrWhiteSpace(tsv[6]) ? (decimal?)null : Convert.ToDecimal(tsv[6], CultureInfo.InvariantCulture);
-            CompanyID = string.IsNullOrWhiteSpace(tsv[7]) ? (int?)null : Convert.ToInt32(tsv[7], CultureInfo.InvariantCulture);
+            CompanyID = string.IsNullOrWhiteSpace(tsv[7]) ? (int?)null : Convert.ToInt64(tsv[7], CultureInfo.InvariantCulture);
             ICBIndustry = string.IsNullOrWhiteSpace(tsv[8]) ? null : tsv[8];
             ICBSuperSector = string.IsNullOrWhiteSpace(tsv[9]) ? null : tsv[9];
             ICBSector = string.IsNullOrWhiteSpace(tsv[10]) ? null : tsv[10];
             ICBSubSector = string.IsNullOrWhiteSpace(tsv[11]) ? null : tsv[11];
-            ICBCode = string.IsNullOrWhiteSpace(tsv[12]) ? (int?)null : Convert.ToInt32(tsv[12], CultureInfo.InvariantCulture);
+            ICBCode = string.IsNullOrWhiteSpace(tsv[12]) ? (int?)null : Convert.ToInt64(tsv[12], CultureInfo.InvariantCulture);
             CompanyName = string.IsNullOrWhiteSpace(tsv[13]) ? null : tsv[13];
             PreviousResultsAnnouncementDate = string.IsNullOrWhiteSpace(tsv[14]) ? (DateTime?)null : DateTime.ParseExact(tsv[14], "yyyyMMdd", CultureInfo.InvariantCulture);
             NextResultsAnnouncementsDate = string.IsNullOrWhiteSpace(tsv[15]) ? (DateTime?)null : DateTime.ParseExact(tsv[15], "yyyyMMdd", CultureInfo.InvariantCulture);
@@ -286,12 +286,12 @@ namespace QuantConnect.DataSource
             LastIDsUpdate = string.IsNullOrWhiteSpace(tsv[indexes[nameof(LastIDsUpdate)]]) ? null : DateTime.ParseExact(tsv[indexes[nameof(LastIDsUpdate)]], "yyyy-MM-dd", CultureInfo.InvariantCulture);
             ISIN = string.IsNullOrWhiteSpace(tsv[indexes[nameof(ISIN)]]) ? null : tsv[indexes[nameof(ISIN)]];
             USDMarketCap = string.IsNullOrWhiteSpace(tsv[indexes[nameof(USDMarketCap)]]) ? null : Convert.ToDecimal(tsv[indexes[nameof(USDMarketCap)]], CultureInfo.InvariantCulture);
-            CompanyID = string.IsNullOrWhiteSpace(tsv[indexes[nameof(CompanyID)]]) ? null : Convert.ToInt32(tsv[indexes[nameof(CompanyID)]], CultureInfo.InvariantCulture);
+            CompanyID = string.IsNullOrWhiteSpace(tsv[indexes[nameof(CompanyID)]]) ? null : Convert.ToInt64(tsv[indexes[nameof(CompanyID)]], CultureInfo.InvariantCulture);
             ICBIndustry = string.IsNullOrWhiteSpace(tsv[indexes[nameof(ICBIndustry)]]) ? null : tsv[indexes[nameof(ICBIndustry)]];
             ICBSuperSector = string.IsNullOrWhiteSpace(tsv[indexes[nameof(ICBSuperSector)]]) ? null : tsv[indexes[nameof(ICBSuperSector)]];
             ICBSector = string.IsNullOrWhiteSpace(tsv[indexes[nameof(ICBSector)]]) ? null : tsv[indexes[nameof(ICBSector)]];
             ICBSubSector = string.IsNullOrWhiteSpace(tsv[indexes[nameof(ICBSubSector)]]) ? null : tsv[indexes[nameof(ICBSubSector)]];
-            ICBCode = string.IsNullOrWhiteSpace(tsv[indexes[nameof(ICBCode)]]) ? null : Convert.ToInt32(tsv[indexes[nameof(ICBCode)]], CultureInfo.InvariantCulture);
+            ICBCode = string.IsNullOrWhiteSpace(tsv[indexes[nameof(ICBCode)]]) ? null : Convert.ToInt64(tsv[indexes[nameof(ICBCode)]], CultureInfo.InvariantCulture);
             CompanyName = string.IsNullOrWhiteSpace(tsv[indexes[nameof(CompanyName)]]) ? null : tsv[indexes[nameof(CompanyName)]];
 
             PreviousResultsAnnouncementDate = string.IsNullOrWhiteSpace(tsv[indexes["previousResultsAnnsDate"]]) ? null : DateTime.ParseExact(tsv[indexes["previousResultsAnnsDate"]], "yyyy-MM-dd", CultureInfo.InvariantCulture);
